# api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ApiResponseBatchPublishResult" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/BatchPublishResult"
          },
          "error" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
      "ApiResponseBatchUnpublishResult" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/BatchUnpublishResult"
          },
          "error" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
      "ApiResponseBillListResponse" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/BillListResponse"
          },
          "error" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
      "ApiResponseBillResponse" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/BillResponse"
          },
          "error" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ApiResponsePaymentInitiationResponse" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/PaymentInitiationResponse"
          },
          "error" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "ApiResponsePreFlightChecklistResponse" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/PreFlightChecklistResponse"
          },
          "error" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
...
      "BatchPublishFailure" : {
        "properties" : {
          "billId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          },
          "unitNumber" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "BatchPublishResult" : {
        "properties" : {
          "failed" : {
            "items" : {
              "$ref" : "#/components/schemas/BatchPublishFailure"
            },
            "type" : "array"
          },
          "succeeded" : {
            "items" : {
              "$ref" : "#/components/schemas/BillResponse"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "BatchUnpublishFailure" : {
        "properties" : {
          "billId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          },
          "unitNumber" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "BatchUnpublishResult" : {
        "properties" : {
          "failed" : {
            "items" : {
              "$ref" : "#/components/schemas/BatchUnpublishFailure"
            },
            "type" : "array"
          },
          "succeeded" : {
            "items" : {
              "$ref" : "#/components/schemas/BillResponse"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
      "BillListResponse" : {
        "properties" : {
          "content" : {
            "items" : {
              "$ref" : "#/components/schemas/BillResponse"
            },
            "type" : "array"
          },
          "metrics" : {
            "$ref" : "#/components/schemas/BillMetricsDTO"
          },
          "number" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "BillMetricsDTO" : {
        "properties" : {
          "pendingDraftsCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "publishedCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalExpectedRevenue" : {
            "type" : "number"
          }
        },
        "type" : "object"
      },
      "BillPropertyBillingMonthRequest" : {
        "properties" : {
          "billingMonth" : {
            "pattern" : "\\d{4}-\\d{2}",
            "type" : "string"
          },
          "propertyId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "required" : [ "billingMonth", "propertyId" ],
        "type" : "object"
      },
      "BillResponse" : {
        "properties" : {
          "billingMonth" : {
            "type" : "string"
          },
          "blockId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "blockName" : {
            "type" : "string"
          },
          "charges" : {
            "items" : {
              "$ref" : "#/components/schemas/ChargeResponse"
            },
            "type" : "array"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "dueDate" : {
            "format" : "date",
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "paidAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "status" : {
            "enum" : [ "PENDING", "PUBLISHED", "PAID", "OVERDUE", "PARTIALLY_PAID" ],
            "type" : "string"
          },
          "tenantName" : {
            "type" : "string"
          },
          "totalAmount" : {
            "type" : "number"
          },
          "unitNumber" : {
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ChargeResponse" : {
        "properties" : {
          "amount" : {
            "type" : "number"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "Pageable" : {
        "properties" : {
          "page" : {
            "format" : "int32",
            "minimum" : 0,
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "minimum" : 1,
            "type" : "integer"
          },
          "sort" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "PaymentInitiationResponse" : {
        "properties" : {
          "amount" : {
            "type" : "number"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "currency" : {
            "type" : "string"
          },
          "gatewayName" : {
            "type" : "string"
          },
          "gatewayTransactionId" : {
            "type" : "string"
          },
          "paymentMethod" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "transactionId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "PreFlightChecklistResponse" : {
        "properties" : {
          "activePayers" : {
            "format" : "int32",
            "type" : "integer"
          },
          "isReady" : {
            "type" : "boolean"
          },
          "meterReadingsEntered" : {
            "format" : "int32",
            "type" : "integer"
          },
          "meterReadingsExpected" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalBeds" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalUnits" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "RecordCashPaymentRequest" : {
        "properties" : {
          "amount" : {
            "type" : "number"
          },
          "note" : {
            "type" : "string"
          },
          "payerUserId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "required" : [ "amount" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "JWT access token from `POST /api/v1/auth/login` or `POST /api/v1/auth/signup`",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "contact" : {
      "email" : "dev@tenantliving.com",
      "name" : "Tenant Living Team"
    },
    "description" : "## Overview\nBackend API for the Tenant Living property management system.\n\n## Authentication\n- Obtain **`accessToken`** and **`refreshToken`** from **`POST /api/v1/auth/login`** or **`POST /api/v1/auth/signup`**.\n- Call protected endpoints with header **`Authorization`** value **`Bearer`** followed by the access token.\n- Rotate access with **`POST /api/v1/auth/refresh`** (send refresh token in JSON body).\n- Introspect a JWT without calling a protected route: **`POST /api/v1/auth/validate`**.\n\n## Errors\nValidation and business errors use **`application/json`** body **`ApiError`** (see schema).\n\n## Roles\n| Role | Description |\n|------|-------------|\n| SUPER_ADMIN | Primary landlord |\n| ADMIN | Co-owner / manager |\n| PROPERTY_STAFF | Operations staff |\n| USER | Tenant (default for self-signup) |\n",
    "license" : {
      "name" : "Proprietary",
      "url" : "https://tenantliving.com"
    },
    "title" : "Tenant Living API",
    "version" : "1.0.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/finance/bills" : {
      "get" : {
        "operationId" : "list_1",
        "parameters" : [ {
          "in" : "query",
          "name" : "propertyId",
          "required" : false,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "billingMonth",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "status",
          "required" : false,
          "schema" : {
            "enum" : [ "PENDING", "PUBLISHED", "PAID", "OVERDUE", "PARTIALLY_PAID" ],
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "search",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "pageable",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/Pageable"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseBillListResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "bill-controller" ]
      }
    },
...
    "/api/v1/finance/bills/batch-publish" : {
      "post" : {
        "operationId" : "batchPublish",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BillPropertyBillingMonthRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseBatchPublishResult"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "bill-controller" ]
      }
    },
    "/api/v1/finance/bills/batch-unpublish" : {
      "post" : {
        "operationId" : "batchUnpublish",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BillPropertyBillingMonthRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseBatchUnpublishResult"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "bill-controller" ]
      }
    },
...
    "/api/v1/finance/bills/pre-flight" : {
      "get" : {
        "operationId" : "getPreFlightChecklist_1",
        "parameters" : [ {
          "in" : "query",
          "name" : "propertyId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "billingMonth",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePreFlightChecklistResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "bill-controller" ]
      }
    },
    "/api/v1/finance/bills/preflight" : {
      "get" : {
        "operationId" : "getPreFlightChecklist",
        "parameters" : [ {
          "in" : "query",
          "name" : "propertyId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "billingMonth",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePreFlightChecklistResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "bill-controller" ]
      }
    },
    "/api/v1/finance/bills/{billId}/cash" : {
      "post" : {
        "operationId" : "recordCashPayment",
        "parameters" : [ {
          "in" : "path",
          "name" : "billId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RecordCashPaymentRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePaymentInitiationResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "bill-controller" ]
      }
    },
...
    "/api/v1/finance/bills/{billId}/online" : {
      "post" : {
        "operationId" : "initiateOnlinePayment",
        "parameters" : [ {
          "in" : "path",
          "name" : "billId",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponsePaymentInitiationResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "bill-controller" ]
      }
    },
    "/api/v1/finance/bills/{id}/mark-paid" : {
      "post" : {
        "operationId" : "markPaid",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseBillResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "bill-controller" ]
      }
    },
    "/api/v1/finance/bills/{id}/publish" : {
      "post" : {
        "operationId" : "publish",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseBillResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "bill-controller" ]
      }
    },
    "/api/v1/finance/bills/{id}/unpublish" : {
      "post" : {
        "operationId" : "unpublish",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseBillResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "bill-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Local (default dev port)",
    "url" : "http://localhost:8080"
  } ]
}